Change nonce count to uint64_t

+++ dyn_miner/dyn_stratum.h
@@ -22,8 +22,8 @@ struct rpc_config_t {
 };
 
 struct stats_t {
-    std::atomic<uint32_t> nonce_count{};
-    std::atomic<uint32_t> share_count{};
+    std::atomic<uint64_t> nonce_count{};
+    std::atomic<uint64_t> share_count{};
     std::atomic<uint32_t> accepted_share_count{};
     std::atomic<uint32_t> rejected_share_count{};
     std::atomic<uint32_t> latest_diff{};
